pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = [
    "Sandenberg Ferreira Melo <77549672+sandenbergmelo@users.noreply.github.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
pwdlib = { extras = ["argon2"], version = "^0.2.0" }
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"
pytest = "^8.3.1"
taskipy = "^1.13.0"
pytest-cov = "^5.0.0"

[tool.taskipy.tasks]
dev = 'fastapi dev fast_zero/app.py'

pre_test = 'task lint'
test = 'pytest --cov=fast_zero -vv'
post_test = 'task coverage'

migrate_upgrade = 'alembic upgrade head'
migrate_generate = 'alembic revision --autogenerate -m'
migrate = 'task migrate_generate'
post_migrate = 'task migrate_upgrade'

coverage = 'coverage html'

lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
